Move example schema to own file.
Features extensive documentation, and fully fleshes out the example schema
to match the one in the spec.

Also updates types so we use Haskell Int32 for GraphQL Int.
