+++ graphql-api.cabal
@@ -73,6 +73,7 @@ test-suite graphql-api-doctests
       ASTTests
       Examples.InputObject
       Examples.UnionExample
+      ExampleSchema
       OrderedMapTests
       ResolverTests
       SchemaTests
@@ -108,6 +109,7 @@ test-suite graphql-api-tests
       Doctests
       Examples.InputObject
       Examples.UnionExample
+      ExampleSchema
       OrderedMapTests
       ResolverTests
       SchemaTests
